provide reference path for model
